upgrade  processConnection
